requirements/project.txt
django==1.3
gondor
BeautifulSoup==3.2.0
mechanize==0.2.5
geopy
lxml
django-piston